// src/app/core/dashboard/branch/model/branchCreationEntity.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface BranchCreationEntity { 
    name?: string;
    address?: string;
    email?: string;
    phone?: string;
    vat?: string;
    type?: BranchCreationEntity.TypeEnum;
    userCode?: string;
}
export namespace BranchCreationEntity {
    export type TypeEnum = 'RESTAURANT' | 'SUPPLIER';
    export const TypeEnum = {
        RESTAURANT: 'RESTAURANT' as TypeEnum,
        SUPPLIER: 'SUPPLIER' as TypeEnum
    };
}